pyproject.toml
[tool.poetry]
name = "diaries-lot-task"
version = "0.1.0"
description = ""
authors = ["Basma Nizar Mohammed Eid <basmanizar@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.5"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
